Remove server management from junit tests
The tests for socket and HTTP servers used to setUp
and tearDown the servers before/after running the tests.
Did not get it to work reliably, so removed this functionality.
The users of "ant test" are required to set up both servers
manually (as was the case before), as documented in the README.